requirements.txt
flask
mysql-connector-python
numpy
gunicorn
opencv-python
deepface